Now roomActions are only triggered manually
Before this commit, switching to a Room using its designated key (eg
[mod+w] would automatically trigger the action associated with it, as
long as the room be empty at the time.

Now, switching to a Room triggers nothing, even if the room is empty
at the time. The action is triggered manually by using [mod+tab] from
within an empty Room. Using this shortcut on a non-empty room still
does what it used to do, aka cycle through all the windows opened in
that room.

The file Rooms.hs, while still very short, does have a certain need
for explanatory comments; anyway I'll mention here that 'goto' is
meant to mean "go to the room without triggering the associated
action, even if the room is empty", which is the new behavior, whereas
"open" is meant to represent the old behavior, which triggers the
action on switching to an empty room.

Motivation for this commit: I kept switching to rooms by mistake,
triggering the spawning of apps which I then had to manually close.

Also, I added a shortcut to cycle through the workspaces: [mod+a].
Motivation: arrow keys are so yesterday, and they're way too far to
the right, which is so wrong.