WIP Add new xmobar at the bottom of screen
The original (top) bar's content is now split between the two bars:
all the indicators except date/time remain where they used to be.
Date/time has been moved to the new (bottom) bar, on the rightside.

The leftside of that new bar contains nothing meaningful for now. I
intend to use that new space to display some real-time info regarding
my music player's state (cmus).

Note that there's also an annoying bug: xmonad doesn't kill one of the
bars on reloading itself, so they keep multiplying everytime i do so.
I'll fix that in one of the commits that are to follow.