Now both bars are at the top of the screen
The one that was at the bottom of the screen bothered me, i kept
hovering over it by mistake, unable to use the scrollwheel.

So now that one is at the very top of the screen, and the one that
used to be at the top alone is now right under the other.

Their names are now pretty confusing obviously, so the next commit
will probably fix that.