Fix restartXMonad: kill xmobars only when logical
Before this commit: If i restarted xmonad, it'd kill the xmobars, try
to recompile, and if successful, would then restart xmonad. Therefore,
if the recompiling failed, i'd be left with no more xmobars even
though xmonad would not even have restarted.

Now: xmobars are killed only when the recompiling is successful.