Now musicbar is started via startupHook
Previously launched via statusBar, which probably spawned it with a
pipe, probably causing some observed semi-crashes which may be related
to unread pipes. Now it should not happen again.

cf wiki.haskell.org/Xmonad/Frequently_asked_questions#…
…XMonad_stops_but_the_current_window_still_responds_to_keys

Next step will be to get the pid of the spawned musicbar so we can
kill it without using the admittedly awkward pkill.